[feature]: Ability to add relative font path css definitions to svg.
When trying to test a component that uses a specific font (e.g. a vector
icon font), it is important to be able to actually see the correct text
in the svg. If the font url is not specified and has not been installed on
the system itself (which fonts like "FontAwesome" are unlikely to be), the
characters display as only an empty box. By specifying the font in the
style tag of the svg, the characters can be displayed properly.